upgraded theme configuration
