+++ .gitignore
@@ -45,4 +45,4 @@ Thumbs.db
 documentation
 
 #Custom Config
-/src/config/custom.ts
+/src/config/custom
