+++ src/app/home/landing/lp-banner/lp-banner.component.ts
@@ -1,4 +1,4 @@
-import { APP_DATA } from './../../../shared/data/app-data';
+import { environment } from './../../../../environments/environment';
 import { Component, OnInit } from '@angular/core';
 
 @Component({
@@ -7,7 +7,7 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./lp-banner.component.scss']
 })
 export class LpBannerComponent implements OnInit {
-  banners = APP_DATA.landing_page_banner;
+  banners = environment.config.landing_page_banner;
   constructor() { }
 
   ngOnInit() {
