+++ src/app/layout/header/components/categories-menu-dropdown/categories-components/categories-details/brand-logo/brand-logo.component.ts
@@ -1,5 +1,4 @@
 import { environment } from './../../../../../../../../environments/environment';
-import { APP_DATA } from './../../../../../../../shared/data/app-data';
 import { Component, OnInit, Input, OnChanges, ChangeDetectionStrategy } from '@angular/core';
 
 @Component({
