+++ src/app/layout/header/components/header-cart/header-cart.component.html
@@ -1,9 +1,9 @@
 <div class="header-cart btn" [routerLink]="['/checkout/cart']" *ngIf="!devicewidth">
   <i class="fa fa-shopping-cart" aria-hidden="true"></i> Cart
-  <span class="badge badge-pill">{{totalCartItems}}</span>
+  <span class="badge badge-pill shadow-sm">{{totalCartItems}}</span>
 </div>
 
 <div class="header-cart btn" [routerLink]="['/checkout/cart']" *ngIf="devicewidth">
   <i class="fa fa-shopping-cart" aria-hidden="true"></i>
-  <span class="badge badge-pill">{{totalCartItems}}</span>
+  <span class="badge badge-pill shadow-sm">{{totalCartItems}}</span>
 </div>
