+++ src/app/layout/header/header.component.scss
@@ -8,6 +8,10 @@
     z-index: 100;
     width: 100%;
   }
+  &:not(#main), &.page-scroll-header {
+    box-shadow: 0 8px 10px -8px $gray-500;
+    z-index: 1;
+  }
 }
 
 .page-scroll-header {
