+++ src/app/shared/scss/custom_themes/custom/_theme_variables.scss
@@ -6,13 +6,14 @@
 $theme-main-color: #a3d032;
 $theme-action-color: #fe9700;
 $primary-font: "Roboto Regular";
+$header-font-color: $white;
 
 $theme-action-sub-color1:#FFB44B;
 $theme-action-sub-color2:#E79110;
 $theme-action-sub-color3: #FFAC2E;
 $theme-action-sub-color4: #E78E0D;
 
-$theme-main-sub-color1:#AAE714; 
+$theme-main-sub-color1:#AAE714;
 $theme-main-sub-color2: #8FB62C;
 $theme-main-sub-color3: #C0E75D;
 $theme-main-sub-color4: #AFDD35
