+++ src/app/shared/scss/custom_themes/default/_header.scss
@@ -1,5 +1,5 @@
 .navbar {
-  color: $white;
+  color: $header-font-color;
   background-color: $theme-main-color;
 
   .header-cart {
