+++ src/app/shared/scss/custom_themes/default/_theme_variables.scss
@@ -3,9 +3,10 @@
 // -- Theme Colors here --
 // Custom theme colors
 // Refer _default_variables to generate colors
-$theme-main-color: #128ced;
-$theme-action-color: #f8d71c;
+$theme-main-color: $gray-200;
+$theme-action-color: #EB7F00 + 100;
 $primary-font: "Whitney";
+$header-font-color: $gray-600;
 
 $theme-action-sub-color1: #EF6C00;
 $theme-action-sub-color2: #FF7300;
