+++ src/config/custom.example/app-data.ts
@@ -1,0 +1,96 @@
+export const APP_DATA = {
+  landing_page_banner: [
+    {
+      image_link: 'assets/custom/banner/page1.jpg',
+      link_url: '#'
+    },
+    {
+      image_link: 'assets/custom/banner/page2.jpg',
+      link_url: '#'
+    },
+    {
+      image_link: 'assets/custom/banner/page3.jpg',
+      link_url: '#'
+    },
+    {
+      image_link: 'assets/custom/banner/page4.jpg',
+      link_url: '#'
+    }
+  ],
+  promo_banner: {
+    image_link: 'assets/custom/banner/promo/promo.png',
+    link_url: '#'
+  },
+  Deals: {
+    type: 'Today\'s deals'
+  },
+
+  Dog: {
+    image_link: 'https://www.what-dog.net/Images/faces2/scroll001.jpg'
+  },
+
+  Cat: {
+    image_link:
+      'https://images.pexels.com/photos/104827/cat-pet-animal-domestic-104827.jpeg?auto=compress&cs=tinysrgb&h=350'
+  },
+
+  footer_page_links: [
+    {
+      name: 'About Us',
+      link_url: '#'
+    },
+    {
+      name: 'Blog',
+      link_url: '#'
+    },
+    {
+      name: 'Return Policy',
+      link_url: '#'
+    },
+    {
+      name: 'FAQs',
+      link_url: '#'
+    },
+    {
+      name: 'Testimonials',
+      link_url: '#'
+    }
+  ],
+  footer_social_links: [
+    {
+      link_url: 'https://twitter.com/angularspree',
+      name: 'Twitter',
+      icon: 'fa fa-twitter-square'
+    },
+    {
+      link_url: 'https://www.instagram.com/angularspree/',
+      name: 'Instagram',
+      icon: 'fa fa-instagram'
+    },
+    {
+      link_url:
+        'https://plus.google.com/b/110371544800340671090/110371544800340671090',
+      name: 'Google +',
+      icon: 'fa fa-google-plus-square'
+    },
+    {
+      link_url: 'https://in.pinterest.com/angularspree/',
+      name: 'Pinterest',
+      icon: 'fa fa-pinterest-square'
+    },
+    {
+      link_url: 'https://www.facebook.com/angularspree/',
+      name: 'Facebook',
+      icon: 'fa fa-facebook-square'
+    },
+    {
+      link_url: 'https://www.youtube.com/channel/UCFBvY3QxKAKBAI_chAzRpjQ',
+      name: 'Youtube',
+      icon: 'fa fa-youtube-square'
+    }
+  ],
+  contact_info: {
+    contact_no: '917-6031-568',
+    copyright: 'Copyright © 2018 angularspree, Inc.'
+  }
+};
